cephadm: move FileLock and friends to locking.py
Pair-programmed-with: Adam King <adking@redhat.com>
